+++ UniWinC/Assets/Kirurobo/UniWindowController/Runtime/Scripts/UniWindowController.cs
@@ -72,9 +72,15 @@ namespace Kirurobo
         }
 
         /// <summary>
+        /// Get the current instance of UniWindowController
+        /// </summary>
+        public static UniWindowController current => _current ? _current : FindOrCreateInstance();
+        private static UniWindowController _current;
+
+        /// <summary>
         /// Low level class
         /// </summary>
-        private UniWinCore uniWinCore = null;
+        private UniWinCore _uniWinCore = null;
 
         /// <summary>
         /// Is this window receives mouse events
@@ -113,7 +119,7 @@ namespace Kirurobo
         /// </summary>
         public bool isTopmost
         {
-            get { return ((uniWinCore == null) ? _isTopmost : _isTopmost = uniWinCore.IsTopmost); }
+            get { return ((_uniWinCore == null) ? _isTopmost : _isTopmost = _uniWinCore.IsTopmost); }
             set { SetTopmost(value); }
         }
         [SerializeField, EditableProperty, Tooltip("Check to set topmost on startup")]
@@ -124,7 +130,7 @@ namespace Kirurobo
         /// </summary>
         public bool isBottommost
         {
-            get { return ((uniWinCore == null) ? _isBottommost : _isBottommost = uniWinCore.IsBottommost); }
+            get { return ((_uniWinCore == null) ? _isBottommost : _isBottommost = _uniWinCore.IsBottommost); }
             set { SetBottommost(value); }
         }
         [SerializeField, EditableProperty, Tooltip("Check to set bottommost on startup")]
@@ -135,7 +141,7 @@ namespace Kirurobo
         /// </summary>
         public bool isZoomed
         {
-            get { return ((uniWinCore == null) ? _isZoomed : _isZoomed = uniWinCore.GetZoomed()); }
+            get { return ((_uniWinCore == null) ? _isZoomed : _isZoomed = _uniWinCore.GetZoomed()); }
             set { SetZoomed(value); }
         }
         [SerializeField, EditableProperty, Tooltip("Check to set zoomed on startup")]
@@ -197,7 +203,6 @@ namespace Kirurobo
         /// </summary>
         private float raycastMaxDepth = 100.0f;
 
-        
         /// <summary>
         /// trueにしておくと、ウィンドウ透過時にカメラ背景を単色の黒透明に自動で変更します
         /// </summary>
@@ -214,6 +219,12 @@ namespace Kirurobo
         /// </summary>
         [Tooltip("Force windowed on startup")]
         public bool forceWindowed = false;
+
+        /// <summary>
+        /// カメラのインスタンス
+        /// </summary>
+        [Tooltip("Main camera is used if None")]
+        public Camera currentCamera;
 
         /// <summary>
         /// 透過方式の指定
@@ -246,8 +257,8 @@ namespace Kirurobo
         /// </summary>
         public Vector2 windowPosition
         {
-            get { return (uniWinCore != null ? uniWinCore.GetWindowPosition() : Vector2.zero); }
-            set { uniWinCore?.SetWindowPosition(value); }
+            get { return (_uniWinCore != null ? _uniWinCore.GetWindowPosition() : Vector2.zero); }
+            set { _uniWinCore?.SetWindowPosition(value); }
         }
 
         /// <summary>
@@ -255,8 +266,8 @@ namespace Kirurobo
         /// </summary>
         public Vector2 windowSize
         {
-            get { return (uniWinCore != null ? uniWinCore.GetWindowSize() : Vector2.zero); }
-            set { uniWinCore?.SetWindowSize(value); }
+            get { return (_uniWinCore != null ? _uniWinCore.GetWindowSize() : Vector2.zero); }
+            set { _uniWinCore?.SetWindowSize(value); }
         }
 
         /// <summary>
@@ -264,8 +275,8 @@ namespace Kirurobo
         /// </summary>
         public Vector2 cursorPosition
         {
-            get { return (uniWinCore != null ? uniWinCore.GetCursorPosition() : Vector2.zero); }
-            set { uniWinCore?.SetCursorPosition(value); }
+            get { return UniWinCore.GetCursorPosition(); }
+            set { UniWinCore.SetCursorPosition(value); }
         }
 
         /// <summary>
@@ -276,11 +287,6 @@ namespace Kirurobo
         // カメラの背景をアルファゼロの黒に置き換えるため、本来の背景を保存しておく変数
         private CameraClearFlags originalCameraClearFlags;
         private Color originalCameraBackground;
-
-        /// <summary>
-        /// カメラのインスタンス
-        /// </summary>
-        private Camera currentCamera;
 
         /// <summary>
         /// カーソル下1px分の色が入るテクスチャ
@@ -320,6 +326,17 @@ namespace Kirurobo
         // Use this for initialization
         void Awake()
         {
+            // シングルトンとする。既にインスタンスがあれば自分を破棄
+            if (this != current)
+            {
+                Destroy(this.gameObject);
+                return;
+            }
+            else
+            {
+                _current = this;
+            }
+
             // フルスクリーン強制解除。エディタでは何もしない
 #if !UNITY_EDITOR
             if (forceWindowed && Screen.fullScreen)
@@ -333,11 +350,11 @@ namespace Kirurobo
                 // メインカメラを探す
                 currentCamera = Camera.main;
 
-                // もしメインカメラが見つからなければ、Findで探す
-                if (!currentCamera)
-                {
-                    currentCamera = GameObject.FindObjectOfType<Camera>();
-                }
+                //// もしメインカメラが見つからなければ、Findで探す
+                //if (!currentCamera)
+                //{
+                //    currentCamera = GameObject.FindObjectOfType<Camera>();
+                //}
             }
 
             // カメラの元の背景を記憶
@@ -358,7 +375,7 @@ namespace Kirurobo
             colorPickerTexture = new Texture2D(1, 1, TextureFormat.ARGB32, false);
 
             // ウィンドウ制御用のインスタンス作成
-            uniWinCore = new UniWinCore();
+            _uniWinCore = new UniWinCore();
         }
 
         /// <summary>
@@ -368,7 +385,7 @@ namespace Kirurobo
         {
             if (!_shouldFitMonitor) return;
 
-            int monitors = uniWinCore.GetMonitorCount();
+            int monitors = UniWinCore.GetMonitorCount();
             int targetMonitorIndex = _monitorToFit;
 
             if (targetMonitorIndex < 0)
@@ -382,8 +399,27 @@ namespace Kirurobo
 
             if (targetMonitorIndex >= 0)
             {
-                uniWinCore.FitToMonitor(targetMonitorIndex);
+                _uniWinCore.FitToMonitor(targetMonitorIndex);
             }
+        }
+
+        /// <summary>
+        /// Find existing instance or create new instance
+        /// </summary>
+        /// <returns></returns>
+        private static UniWindowController FindOrCreateInstance()
+        {
+            var instance = GameObject.FindObjectOfType<UniWindowController>();
+            
+            // 勝手に生成するのは今のところ無効としてみる
+            // // シーンに見つからなければ新規作成
+            // if (!instance)
+            // {
+            //     var obj = new GameObject(nameof(UniWindowController));
+            //     obj.AddComponent<UniWindowController>();
+            // }
+
+            return instance;
         }
         
         void Start()
@@ -401,18 +437,24 @@ namespace Kirurobo
 
         void OnDestroy()
         {
-            if (uniWinCore != null)
+            if (_uniWinCore != null)
             {
-                uniWinCore.Dispose();
+                _uniWinCore.Dispose();
+            }
+
+            // Instance も破棄
+            if (this == current)
+            {
+                _current = null;
             }
         }
 
         void StoreOriginalWindowRectangle()
         {
-            if (uniWinCore != null)
+            if (_uniWinCore != null)
             {
-                var size = uniWinCore.GetWindowSize();
-                var pos = uniWinCore.GetWindowPosition();
+                var size = _uniWinCore.GetWindowSize();
+                var pos = _uniWinCore.GetWindowPosition();
                 originalWindowRectangle = new Rect(pos, size);
             }
         }
@@ -421,12 +463,12 @@ namespace Kirurobo
         void Update()
         {
             // 自ウィンドウ取得ができていなければ、取得
-            if (uniWinCore == null || !uniWinCore.IsActive)
+            if (_uniWinCore == null || !_uniWinCore.IsActive)
             {
                 UpdateTargetWindow();
             } else
             {
-                uniWinCore.Update();
+                _uniWinCore.Update();
             }
             
             // Process events
@@ -441,19 +483,19 @@ namespace Kirurobo
         /// </summary>
         private void UpdateEvents()
         {
-            if (uniWinCore == null) return;
+            if (_uniWinCore == null) return;
 
-            if (uniWinCore.ObserveDroppedFiles(out var droppedFiles))
+            if (_uniWinCore.ObserveDroppedFiles(out var droppedFiles))
             {
                 OnDropFiles?.Invoke(droppedFiles);
             }
 
-            if (uniWinCore.ObserveMonitorChanged())
+            if (_uniWinCore.ObserveMonitorChanged())
             {
                 OnMonitorChanged?.Invoke();
             }
 
-            if (uniWinCore.ObserveWindowStyleChanged(out var type))
+            if (_uniWinCore.ObserveWindowStyleChanged(out var type))
             {
                 // // モニタへのフィット指定がある状態で最大化解除された場合
                 // if (shouldFitMonitor && !uniWinCore.GetZoomed())
@@ -471,8 +513,32 @@ namespace Kirurobo
         IEnumerator ForceZoomed()
         {
             yield return new WaitForSeconds(0.5f);
-            if (_shouldFitMonitor && !uniWinCore.GetZoomed()) SetZoomed(true);
+            if (_shouldFitMonitor && !_uniWinCore.GetZoomed()) SetZoomed(true);
             yield return null;
+        }
+
+        /// <summary>
+        /// カメラを指定。以前のカメラがあれば背景を戻す
+        /// </summary>
+        /// <param name="newCamera"></param>
+        public void SetCamera(Camera newCamera)
+        {
+            // カメラが変更された場合、設定を戻す
+            if (newCamera != currentCamera)
+            {
+                SetCameraBackground(false);
+            }
+
+            currentCamera = newCamera;
+
+            // カメラの元の背景を記憶
+            if (currentCamera)
+            {
+                originalCameraClearFlags = currentCamera.clearFlags;
+                originalCameraBackground = currentCamera.backgroundColor;
+
+                SetCameraBackground(_isTransparent);
+            }
         }
 
         /// <summary>
@@ -481,7 +547,7 @@ namespace Kirurobo
         /// <param name="isThrough"></param>
         void SetClickThrough(bool isThrough)
         {
-            uniWinCore?.EnableClickThrough(isThrough);
+            _uniWinCore?.EnableClickThrough(isThrough);
             _isClickThrough = isThrough;
         }
 
@@ -642,21 +708,29 @@ namespace Kirurobo
             //     return;
             // }
 
-            // 3Dオブジェクトの上か否かを判定
-            Ray ray = currentCamera.ScreenPointToRay(position);
-            if (Physics.Raycast(ray, out _, raycastMaxDepth))
+            if (currentCamera && currentCamera.isActiveAndEnabled)
             {
-                onObject = true;
-                return;
-            }
+                Ray ray = currentCamera.ScreenPointToRay(position);
 
-            // 2Dオブジェクトの上か判定
-            var rayHit2D = Physics2D.GetRayIntersection(ray);
-            Debug.DrawRay(ray.origin, ray.direction, Color.blue, 2f, false);
-            if (rayHit2D.collider != null)
+                // 3Dオブジェクトの上か否かを判定
+                if (Physics.Raycast(ray, out _, raycastMaxDepth))
+                {
+                    onObject = true;
+                    return;
+                }
+
+                // 2Dオブジェクトの上か判定
+                var rayHit2D = Physics2D.GetRayIntersection(ray);
+                Debug.DrawRay(ray.origin, ray.direction, Color.blue, 2f, false);
+                if (rayHit2D.collider != null)
+                {
+                    onObject = true;
+                    return;
+                }
+            } else
             {
-                onObject = true;
-                return;
+                // カメラが有効でなければメインカメラを取得
+                currentCamera = Camera.main;
             }
 
             // いずれもヒットしなければオブジェクト上ではないと判断
@@ -668,22 +742,22 @@ namespace Kirurobo
         /// </summary>
         private void UpdateTargetWindow()
         {
-            if (uniWinCore == null)
+            if (_uniWinCore == null)
             {
-                uniWinCore = new UniWinCore();
+                _uniWinCore = new UniWinCore();
             }
 
             // ウィンドウがまだ取得できていなければ、取得の処理を行う
-            if (!uniWinCore.IsActive)
+            if (!_uniWinCore.IsActive)
             {
-                uniWinCore.AttachMyWindow();
+                _uniWinCore.AttachMyWindow();
 
                 // ウィンドウを取得できたら最初の値を設定
-                if (uniWinCore.IsActive)
+                if (_uniWinCore.IsActive)
                 {
-                    uniWinCore.SetTransparentType((UniWinCore.TransparentType)transparentType);
-                    uniWinCore.SetKeyColor(keyColor);
-                    uniWinCore.SetAlphaValue(_alphaValue);
+                    _uniWinCore.SetTransparentType((UniWinCore.TransparentType)transparentType);
+                    _uniWinCore.SetKeyColor(keyColor);
+                    _uniWinCore.SetAlphaValue(_alphaValue);
                     SetTransparent(_isTransparent);
                     if (_isBottommost)
                     {
@@ -706,7 +780,7 @@ namespace Kirurobo
                 #if UNITY_EDITOR
                 // エディタではゲームビューが閉じられたりドッキングされたりするため、変化していれば対象ウィンドウを変更
                 // アクティブウィンドウが現在の対象と同じならばなにもおこらない
-                uniWinCore.AttachMyActiveWindow();
+                _uniWinCore.AttachMyActiveWindow();
                 #endif
             }
         }
@@ -769,9 +843,9 @@ namespace Kirurobo
             _isTransparent = transparent;
             SetCameraBackground(transparent);
 #if !UNITY_EDITOR
-            if (uniWinCore != null)
+            if (_uniWinCore != null)
             {
-                uniWinCore.EnableTransparent(transparent);
+                _uniWinCore.EnableTransparent(transparent);
             }
 #endif
             UpdateClickThrough();
@@ -783,18 +857,18 @@ namespace Kirurobo
         /// <param name="type"></param>
         public void SetTransparentType(TransparentType type)
         {
-            if (uniWinCore != null) {
+            if (_uniWinCore != null) {
                 // 透過中だったなら、一度解除して再透過
                 if (_isTransparent)
                 {
                     SetTransparent(false);
-                    uniWinCore.SetTransparentType((UniWinCore.TransparentType)type);
+                    _uniWinCore.SetTransparentType((UniWinCore.TransparentType)type);
                     transparentType = type;
                     SetTransparent(true);
                 }
                 else
                 {
-                    uniWinCore.SetTransparentType((UniWinCore.TransparentType)type);
+                    _uniWinCore.SetTransparentType((UniWinCore.TransparentType)type);
                     transparentType = type;
                 }
             }
@@ -807,7 +881,7 @@ namespace Kirurobo
         private void SetAlphaValue(float alpha)
         {
             _alphaValue = alpha;
-            uniWinCore?.SetAlphaValue(_alphaValue);
+            _uniWinCore?.SetAlphaValue(_alphaValue);
         }
 
         /// <summary>
@@ -817,11 +891,11 @@ namespace Kirurobo
         private void SetTopmost(bool topmost)
         {
             //if (_isTopmost == topmost) return;
-            if (uniWinCore == null) return;
+            if (_uniWinCore == null) return;
 
-            uniWinCore.EnableTopmost(topmost);
-            _isTopmost = uniWinCore.IsTopmost;
-            _isBottommost = uniWinCore.IsBottommost;
+            _uniWinCore.EnableTopmost(topmost);
+            _isTopmost = _uniWinCore.IsTopmost;
+            _isBottommost = _uniWinCore.IsBottommost;
         }
 
         /// <summary>
@@ -830,11 +904,11 @@ namespace Kirurobo
         /// <param name="bottommost"></param>
         private void SetBottommost(bool bottommost)
         {
-            if (uniWinCore == null) return;
+            if (_uniWinCore == null) return;
 
-            uniWinCore.EnableBottommost(bottommost);
-            _isBottommost = uniWinCore.IsBottommost;
-            _isTopmost = uniWinCore.IsTopmost;
+            _uniWinCore.EnableBottommost(bottommost);
+            _isBottommost = _uniWinCore.IsBottommost;
+            _isTopmost = _uniWinCore.IsTopmost;
         }
 
         /// <summary>
@@ -843,29 +917,43 @@ namespace Kirurobo
         /// <param name="zoomed"></param>
         private void SetZoomed(bool zoomed)
         {
-            if (uniWinCore == null) return;
+            if (_uniWinCore == null) return;
 
-            uniWinCore.SetZoomed(zoomed);
-            _isZoomed = uniWinCore.GetZoomed();
+            _uniWinCore.SetZoomed(zoomed);
+            _isZoomed = _uniWinCore.GetZoomed();
         }
 
         private void SetAllowDrop(bool enabled)
         {
-            if (uniWinCore == null) return;
+            if (_uniWinCore == null) return;
 
-            uniWinCore.SetAllowDrop(enabled);
+            _uniWinCore.SetAllowDrop(enabled);
             _allowDropFiles = enabled;
         }
 
 
         /// <summary>
-        /// 接続されているモニタ数を取得
+        /// Get the number of connected monitors
         /// </summary>
         /// <returns></returns>
-        public int GetMonitorCount()
+        public static int GetMonitorCount()
         {
-            if (uniWinCore == null) return 0;
-            return uniWinCore.GetMonitorCount();
+            //if (uniWinCore == null) return 0;
+            return UniWinCore.GetMonitorCount();
+        }
+
+        /// <summary>
+        /// Get monitor position and size
+        /// </summary>
+        /// <param name="index"></param>
+        /// <returns></returns>
+        public static Rect GetMonitorRect(int index)
+        {
+            if (UniWinCore.GetMonitorRectangle(index, out Vector2 position, out Vector2 size))
+            {
+                return new Rect(position, size);
+            }
+            return Rect.zero;
         }
 
         /// <summary>
@@ -874,7 +962,7 @@ namespace Kirurobo
         /// <returns></returns>
         private bool FitToMonitor(bool shouldFit, int monitorIndex)
         {
-            if (uniWinCore == null)
+            if (_uniWinCore == null)
             {
                 _shouldFitMonitor = shouldFit;
                 _monitorToFit = monitorIndex;
@@ -909,7 +997,7 @@ namespace Kirurobo
                     _shouldFitMonitor = shouldFit;
                     UpdateMonitorFitting();
 
-                    uniWinCore.SetZoomed(false);
+                    _uniWinCore.SetZoomed(false);
                     //uniWinCore.SetWindowSize(originalWindowRectangle.size);
                     //uniWinCore.SetWindowPosition(originalWindowRectangle.position);
                 }
@@ -924,23 +1012,41 @@ namespace Kirurobo
         }
 
         /// <summary>
+        /// Get mouse cursor position
+        /// </summary>
+        /// <returns>Cursor position</returns>
+        public static Vector2 GetCursorPosition()
+        {
+            return UniWinCore.GetCursorPosition();
+        }
+        
+        /// <summary>
+        /// Set mouse cursor position
+        /// </summary>
+        /// <param name="position"></param>
+        public static void SetCursorPosition(Vector2 position)
+        {
+            UniWinCore.SetCursorPosition(position);
+        }
+
+        /// <summary>
         /// 終了時にはウィンドウ状態を戻す処理が必要
         /// </summary>
         void OnApplicationQuit()
         {
             if (Application.isPlaying)
             {
-                if (uniWinCore != null)
+                if (_uniWinCore != null)
                 {
                     // エディタだとウィンドウ状態を戻す
                     // スタンドアローンだと戻した姿が見えてしまうためスキップ
 #if UNITY_EDITOR
-                    uniWinCore.SetWindowSize(originalWindowRectangle.size);
-                    uniWinCore.SetWindowPosition(originalWindowRectangle.position);
+                    _uniWinCore.SetWindowSize(originalWindowRectangle.size);
+                    _uniWinCore.SetWindowPosition(originalWindowRectangle.position);
 
-                    uniWinCore.DetachWindow();
+                    _uniWinCore.DetachWindow();
 #endif
-                    uniWinCore.Dispose();
+                    _uniWinCore.Dispose();
                 }
             }
         }
@@ -950,7 +1056,7 @@ namespace Kirurobo
         /// </summary>
         public void Focus()
         {
-            if (uniWinCore != null)
+            if (_uniWinCore != null)
             {
                 //uniWin.SetFocus();
             }
@@ -964,7 +1070,7 @@ namespace Kirurobo
         [Obsolete]
         public int GetDebugInfo()
         {
-            if (uniWinCore != null) {
+            if (_uniWinCore != null) {
                 return UniWinCore.GetDebugInfo();
             }
             return 0;