+++ UniWinC/Assets/Kirurobo/UniWindowController/CHANGELOG.md
@@ -8,6 +8,18 @@ How to write the changelog.
 https://keepachangelog.com/ja/1.0.0/
 --->
 
+## [v0.8.4] - 2021-11-27
+### Changed
+- Made the class singleton.
+- All samples are bundled for package manager.
+
+### Added
+- File type selection in macOS
+
+## [v0.8.3] - 2021-11-27
+### Added
+- SetAlphaValue
+
 ## [v0.8.2] - 2021-10-15
 ### Added
 - FilePanel.OpenFilePanel()